+++ lib/recipe/recipesdinner.dart
@@ -99,7 +99,7 @@ class Recipesdinner extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    if (BMICalculator.calculateBMI(double.parse(myController.weightController.text), double.parse(myController.heightController.text), int.parse(myController.ageController.text)) >= 25) {
+    if (BMICalculator.calculateBMI(int.parse(myController.weightController.text), int.parse(myController.heightController.text), int.parse(myController.ageController.text)) >= 25) {
       _dishes.clear();
       _dishes.addAll([
         Dishes(
@@ -177,7 +177,7 @@ class Recipesdinner extends StatelessWidget {
           cooktime: '40 мин',
         ),
       ]);
-    } else if (BMICalculator.calculateBMI(double.parse(myController.weightController.text), double.parse(myController.heightController.text), int.parse(myController.ageController.text)) < 18.6) {
+    } else if (BMICalculator.calculateBMI(int.parse(myController.weightController.text), int.parse(myController.heightController.text), int.parse(myController.ageController.text)) < 18.6) {
       _dishes.clear();
       _dishes.addAll([
         Dishes(
